+++ ydb/public/api/grpc/draft/ya.make
@@ -21,6 +21,7 @@ SRCS(
     ydb_replication_v1.proto
     ydb_sqs_topic_v1.proto
     ydb_tablet_v1.proto
+    ydb_topic_deferred_publish_v1.proto
     ydb_view_v1.proto
     ydb_ymq_v1.proto
 )
