+++ ydb/public/api/grpc/draft/ydb_topic_deferred_publish_v1.proto
@@ -5,7 +5,7 @@ package Ydb.Topic.DeferredPublish.V1;
 
 option java_package = "com.yandex.ydb.topic.deferredpublish.v1";
 
-import "ydb/public/api/protos/ydb_topic_deferred_publish.proto";
+import "ydb/public/api/protos/draft/ydb_topic_deferred_publish.proto";
 
 service TopicDeferredPublishService {
     // Starts a deferred topic publication.
