+++ ydb/public/api/grpc/ya.make
@@ -26,7 +26,6 @@ SRCS(
     ydb_scripting_v1.proto
     ydb_table_v1.proto
     ydb_test_shard_v1.proto
-    ydb_topic_deferred_publish_v1.proto
     ydb_topic_v1.proto
 )
