+++ ydb/public/api/protos/ya.make
@@ -22,6 +22,7 @@ SRCS(
     draft/ydb_object_storage.proto
     draft/ydb_replication.proto
     draft/ydb_tablet.proto
+    draft/ydb_topic_deferred_publish.proto
     draft/ydb_view.proto
     draft/ymq.proto
     persqueue_error_codes_v1.proto
@@ -54,7 +55,6 @@ SRCS(
     ydb_table.proto
     ydb_test_shard.proto
     ydb_topic.proto
-    ydb_topic_deferred_publish.proto
     ydb_value.proto
 )
