+++ ydb/services/persqueue_v1/ut/ya.make
@@ -53,6 +53,7 @@ PEERDIR(
     ydb/library/persqueue/tests
     ydb/library/persqueue/topic_parser
     ydb/public/api/grpc
+    ydb/public/api/grpc/draft
     ydb/public/sdk/cpp/src/client/persqueue_public/ut/ut_utils
     ydb/public/sdk/cpp/src/client/persqueue_public
     ydb/public/sdk/cpp/src/client/table
